avoid infinite recursion in HttpMessageBus constructor
We had a case of infintie recursion in HTTPMessageBus constructor because it initialised an actor with a default bus (namely a TCPServerActor). This actor then requested the default bus, which wasn't set yet, so was created. Recurse ad nauseum.

Also better error handling in HTTPMessageBus when dealing with bad requests