Implement Actor.await(Callable<T> closure) -> T.
The new method Actor.await(Callable) runs a Callable and returns the return value of the callable. The trick is that this methods informs the underlying thread pool that we will do blocking ops on the the running thread, allowing the pool to dispatch an extra worker in the mean time (if applicable). This is very useful for IO or other blocking ops.